backend/requirements.txt
flask
flask-socketio
kafka-python
eventlet          # Enables async sockets for Flask‑SocketIO

